requirements.txt

# Flask Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Environment Variables
python-dotenv==1.0.0

# Machine Learning & NLP
transformers==4.35.2
torch==2.0.1
scikit-learn==1.3.2
numpy==1.24.3
nltk==3.8.1
datasets==2.14.6
evaluate==0.4.1
accelerate==0.24.1
tokenizers==0.15.0

# Explainability
lime==0.2.0.1

# Translation
deep-translator==1.11.4

# HTTP Requests
requests==2.31.0

